Initial version
My custom dictionary is about 3 times slower than the stock Dictionary in the BCL. This is largely caused by the fact that we use a fairly small bucket size by default, and we resize every time we need it. This degrades performance significantly; if we use a larger bucket size right away, it makes us perform only about 50% slower than the built-in implementation. :)